Move error code enums to separate files
