+++ saleor/graphql/account/i18n.py
@@ -1,10 +1,10 @@
 from django.core.exceptions import ValidationError
 from django_countries import countries
 
+from ...account.error_codes import AccountErrorCode
 from ...account.forms import get_address_form
 from ...account.models import Address
 from ...account.validators import validate_possible_number
-from ..core.utils.error_codes import AccountErrorCode
 
 
 class I18nMixin:
