+++ saleor/graphql/account/mutations/base.py
@@ -7,6 +7,7 @@ from graphql_jwt.exceptions import PermissionDenied
 
 from ....account import events as account_events, models
 from ....account.emails import send_user_password_reset_email_with_url
+from ....account.error_codes import AccountErrorCode
 from ....core.utils.url import validate_storefront_url
 from ....dashboard.emails import send_set_password_email_with_url
 from ...account.i18n import I18nMixin
@@ -21,7 +22,6 @@ from ...core.mutations import (
     validation_error_to_error_type,
 )
 from ...core.types.common import AccountError
-from ...core.utils.error_codes import AccountErrorCode
 
 BILLING_ADDRESS_FIELD = "default_billing_address"
 SHIPPING_ADDRESS_FIELD = "default_shipping_address"
