+++ saleor/graphql/account/mutations/deprecated_account.py
@@ -2,11 +2,11 @@ import graphene
 from django.core.exceptions import ObjectDoesNotExist, ValidationError
 
 from ....account import events as account_events, models, utils
+from ....account.error_codes import AccountErrorCode
 from ....checkout import AddressType
 from ...account.enums import AddressTypeEnum
 from ...account.types import Address, AddressInput, User
 from ...core.mutations import BaseMutation, ModelMutation
-from ...core.utils.error_codes import AccountErrorCode
 from .base import BaseCustomerCreate, UserAddressInput
 from .deprecated_staff import send_user_password_reset_email
