+++ saleor/graphql/core/utils/error_codes.py
@@ -1,5 +1,16 @@
 from enum import Enum
 
+from ....account.error_codes import AccountErrorCode
+from ....checkout.error_codes import CheckoutErrorCode
+from ....core.error_codes import ShopErrorCode
+from ....extensions.error_codes import ExtensionsErrorCode
+from ....giftcard.error_codes import GiftcardErrorCode
+from ....menu.error_codes import MenuErrorCode
+from ....order.error_codes import OrderErrorCode
+from ....payment.error_codes import PaymentErrorCode
+from ....product.error_codes import ProductErrorCode
+from ....shipping.error_codes import ShippingErrorCode
+
 DJANGO_VALIDATORS_ERROR_CODES = [
     "invalid",
     "invalid_extension",
@@ -26,146 +37,6 @@ DJANGO_FORM_FIELDS_ERROR_CODES = [
     "missing",
     "overflow",
 ]
-
-
-class AccountErrorCode(Enum):
-    ACTIVATE_OWN_ACCOUNT = "activate_own_account"
-    ACTIVATE_SUPERUSER_ACCOUNT = "activate_superuser_account"
-    DEACTIVATE_OWN_ACCOUNT = "deactivate_own_account"
-    DEACTIVATE_SUPERUSER_ACCOUNT = "deactivate_superuser_account"
-    DELETE_NON_STAFF_USER = "delete_non_staff_user"
-    DELETE_OWN_ACCOUNT = "delete_own_account"
-    DELETE_STAFF_ACCOUNT = "delete_staff_account"
-    DELETE_SUPERUSER_ACCOUNT = "delete_superuser_account"
-    GRAPHQL_ERROR = "graphql_error"
-    INVALID = "invalid"
-    INVALID_PASSWORD = "invalid_password"
-    NOT_FOUND = "not_found"
-    PASSWORD_ENTIRELY_NUMERIC = "password_entirely_numeric"
-    PASSWORD_TOO_COMMON = "password_too_common"
-    PASSWORD_TOO_SHORT = "password_too_short"
-    PASSWORD_TOO_SIMILAR = "password_too_similar"
-    REQUIRED = "required"
-    UNIQUE = "unique"
-
-
-class CheckoutErrorCode(Enum):
-    BILLING_ADDRESS_NOT_SET = "billing_address_not_set"
-    CHECKOUT_NOT_FULLY_PAID = "checkout_not_fully_paid"
-    GRAPHQL_ERROR = "graphql_error"
-    INSUFFICIENT_STOCK = "insufficient_stock"
-    INVALID = "invalid"
-    INVALID_SHIPPING_METHOD = "invalid_shipping_method"
-    NOT_FOUND = "not_found"
-    PAYMENT_ERROR = "payment_error"
-    QUANTITY_GREATER_THAN_LIMIT = "quantity_greater_than_limit"
-    REQUIRED = "required"
-    SHIPPING_ADDRESS_NOT_SET = "shipping_address_not_set"
-    SHIPPING_METHOD_NOT_APPLICABLE = "shipping_method_not_applicable"
-    SHIPPING_METHOD_NOT_SET = "shipping_method_not_set"
-    SHIPPING_NOT_REQUIRED = "shipping_not_required"
-    TAX_ERROR = "tax_error"
-    UNIQUE = "unique"
-    VOUCHER_NOT_APPLICABLE = "voucher_not_applicable"
-    ZERO_QUANTITY = "zero_quantity"
-
-
-class MenuErrorCode(Enum):
-    CANNOT_ASSIGN_NODE = "cannot_assign_node"
-    GRAPHQL_ERROR = "graphql_error"
-    INVALID = "invalid"
-    INVALID_MENU_ITEM = "invalid_menu_item"
-    NO_MENU_ITEM_PROVIDED = "no_item_provided"
-    NOT_FOUND = "not_found"
-    REQUIRED = "required"
-    TOO_MANY_MENU_ITEMS = "too_many_items"
-    UNIQUE = "unique"
-
-
-class OrderErrorCode(Enum):
-    CANNOT_CANCEL_FULFILLMENT = "cannot_cancel_fulfillment"
-    CANNOT_CANCEL_ORDER = "cannot_cancel_order"
-    CANNOT_DELETE = "cannot_delete"
-    CANNOT_REFUND = "cannot_refund"
-    CAPTURE_INACTIVE_PAYMENT = "capture_inactive_payment"
-    EDIT_NON_DRAFT_ORDER = "edit_non_draft_order"
-    FULFILL_ORDER_LINE = "fulfill_order_line"
-    GRAPHQL_ERROR = "graphql_error"
-    INVALID = "invalid"
-    NOT_FOUND = "not_found"
-    ORDER_INVALID_SHIPPING_METHOD = "order_invalid_shipping_method"
-    ORDER_NO_SHIPPING_ADDRESS = "order_no_shipping_address"
-    PAYMENT_ERROR = "payment_error"
-    PAYMENT_MISSING = "payment_missing"
-    REQUIRED = "required"
-    SHIPPING_METHOD_REQUIRED = "shipping_method_required"
-    UNIQUE = "unique"
-    VOID_INACTIVE_PAYMENT = "void_inactive_payment"
-    ZERO_QUANTITY = "zero_quantity"
-
-
-class ProductErrorCode(Enum):
-    ALREADY_EXISTS = "already_exists"
-    ATTRIBUTE_ALREADY_ASSIGNED = "attribute_already_assigned"
-    ATTRIBUTE_BAD_VALUE = "attribute_bad_value"
-    ATTRIBUTE_CANNOT_BE_ASSIGNED = "attribute_cannot_be_assigned"
-    ATTRIBUTE_DISABLED_VARIANTS = "attribute_disabled_variants"
-    GRAPHQL_ERROR = "graphql_error"
-    INVALID = "invalid"
-    NOT_PRODUCTS_IMAGE = "not_products_image"
-    NOT_FOUND = "not_found"
-    REQUIRED = "required"
-    UNIQUE = "unique"
-    VARIANT_NO_DIGITAL_CONTENT = "variant_no_digital_content"
-
-
-class ShopErrorCode(Enum):
-    ALREADY_EXISTS = "already_exists"
-    CANNOT_FETCH_TAX_RATES = "cannot_fetch_tax_rates"
-    GRAPHQL_ERROR = "graphql_error"
-    INVALID = "invalid"
-    NOT_FOUND = "not_found"
-    REQUIRED = "required"
-    UNIQUE = "unique"
-
-
-class ShippingErrorCode(Enum):
-    ALREADY_EXISTS = "already_exists"
-    GRAPHQL_ERROR = "graphql_error"
-    INVALID = "invalid"
-    MAX_LESS_THAN_MIN = "max_less_than_min"
-    NOT_FOUND = "not_found"
-    REQUIRED = "required"
-    UNIQUE = "unique"
-
-
-class PaymentErrorCode(Enum):
-    BILLING_ADDRESS_NOT_SET = "billing_address_not_set"
-    GRAPHQL_ERROR = "graphql_error"
-    INVALID = "invalid"
-    NOT_FOUND = "not_found"
-    PARTIAL_PAYMENT_NOT_ALLOWED = "partial_payment_not_allowed"
-    PAYMENT_ERROR = "payment_error"
-    REQUIRED = "required"
-    UNIQUE = "unique"
-
-
-class GiftcardErrorCode(Enum):
-    ALREADY_EXISTS = "already_exists"
-    GRAPHQL_ERROR = "graphql_error"
-    INVALID = "invalid"
-    NOT_FOUND = "not_found"
-    REQUIRED = "required"
-    UNIQUE = "unique"
-
-
-class ExtensionsErrorCode(Enum):
-    GRAPHQL_ERROR = "graphql_error"
-    INVALID = "invalid"
-    PLUGIN_MISCONFIGURED = "plugin-misconfigured"
-    NOT_FOUND = "not_found"
-    REQUIRED = "required"
-    UNIQUE = "unique"
 
 
 SALEOR_ERROR_CODE_ENUMS = [
