+++ saleor/shipping/error_codes.py
@@ -1,0 +1,11 @@
+from enum import Enum
+
+
+class ShippingErrorCode(Enum):
+    ALREADY_EXISTS = "already_exists"
+    GRAPHQL_ERROR = "graphql_error"
+    INVALID = "invalid"
+    MAX_LESS_THAN_MIN = "max_less_than_min"
+    NOT_FOUND = "not_found"
+    REQUIRED = "required"
+    UNIQUE = "unique"
