+++ tests/api/test_digital_content.py
@@ -1,6 +1,6 @@
 import graphene
 
-from saleor.graphql.core.utils.error_codes import ProductErrorCode
+from saleor.product.error_codes import ProductErrorCode
 from saleor.product.models import DigitalContent, ProductVariant
 from tests.api.utils import get_graphql_content
 from tests.utils import create_image
